style: apply linter fix

+++ pkg/decryption/client.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Client interface {
-	DecryptionReader(ctx context.Context, name string, size int, r io.Reader) (io.Reader, error)
+	DecryptionReader(ctx context.Context, name string, size uint64, r io.Reader) (io.Reader, error)
 }
 
 type client struct{}
@@ -18,7 +18,7 @@ func New() Client {
 	return &client{}
 }
 
-func (*client) DecryptionReader(ctx context.Context, name string, size int, r io.Reader) (io.Reader, error) {
+func (*client) DecryptionReader(ctx context.Context, name string, size uint64, r io.Reader) (io.Reader, error) {
 	slog.DebugContext(ctx, "Setting up decryption reader", "name", name)
 	decryptor := decryptionReader(FileFormat(filepath.Ext(name)))
 	if decryptor == nil {
